+++ dictionary/mix.exs
@@ -14,7 +14,7 @@ defmodule Dictionary.Mixfile do
   def application do
     [
       extra_applications: [:logger],
-      mod: { Dictionary , [] }
+      mod: { Dictionary.Application , [] }
     ]
   end
