+++ hangman/lib/hangman/server.ex
@@ -1,30 +1,12 @@
 defmodule Hangman.Server do
 
   alias Hangman.Game, as: Game
+
   use GenServer
-  @name Hangman.Server
 
   def start_link() do
-    GenServer.start_link(@name, [], name: @name)
+    GenServer.start_link(__MODULE__, [], name: __MODULE__)
   end
-
-  def new_game() do
-    GenServer.call(@name, :new_game)
-  end
-
-
-  def new_game(word) do
-    GenServer.call(@name, { :new_game, word })
-  end
-
-  def make_move(game, guess) do
-    GenServer.call(@name, { :make_move, game, guess })
-  end
-
-  def tally(game) do
-    GenServer.call(@name, { :tally, game })
-  end
-
 
   ######Implementation######
   def handle_call(:new_game, _from, state) do
@@ -42,6 +24,5 @@ defmodule Hangman.Server do
   def handle_call({ :tally, game }, _from, state) do
     { :reply, Game.tally(game), state }
   end
-
-
+  
 end
