+++ hangman/lib/hangman/state.ex
@@ -7,5 +7,6 @@ defmodule Hangman.State do
     turns_left: 7,
     used:       MapSet.new([])
   )
+
 end
