+++ hangman/lib/hangman/supervisor.ex
@@ -1,0 +1,17 @@
+defmodule Hangman.Supervisor do
+
+  use Supervisor
+
+  def start_link() do
+    Supervisor.start_link(__MODULE__, [], name: __MODULE__)
+  end
+
+  def init(_arg) do
+    children = [
+      worker(Hangman.Server, [], restart: :transient)
+    ]
+
+    supervise(children, strategy: :one_for_one)
+  end
+
+end
